+++ packages/api-v4/src/account/types.ts
@@ -14,7 +14,7 @@ export interface Account {
   email: string;
   first_name: string;
   tax_id: string;
-  credit_card: CreditCard;
+  credit_card: CreditCardData;
   state: string;
   zip: string;
   address_1: string;
@@ -38,7 +38,8 @@ export type AccountCapability =
   | 'Kubernetes'
   | 'Cloud Firewall'
   | 'Vlans'
-  | 'Machine Images';
+  | 'Machine Images'
+  | 'LKE HA Control Planes';
 
 export interface AccountSettings {
   managed: boolean;
@@ -84,12 +85,6 @@ export type CardType =
   | 'JCB';
 
 export type PaymentType = 'credit_card' | ThirdPartyPayment;
-
-export interface CreditCard {
-  expiry: string | null;
-  last_four: string | null;
-  card_type?: CardType;
-}
 
 export interface Invoice {
   id: number;
@@ -395,19 +390,38 @@ export interface AccountMaintenance {
   };
 }
 
-export interface PaymentMethod {
-  id: number;
-  type: PaymentType;
-  is_default: boolean;
-  created: string;
-  data: CreditCard;
+export interface PayPalData {
+  paypal_id: string;
+  email: string;
 }
+
+export interface CreditCardData {
+  expiry: string | null;
+  last_four: string | null;
+  card_type?: CardType;
+}
+
+export type PaymentMethod =
+  | {
+      id: number;
+      type: 'credit_card' | 'google_pay';
+      is_default: boolean;
+      created: string;
+      data: CreditCardData;
+    }
+  | {
+      id: number;
+      type: 'paypal';
+      is_default: boolean;
+      created: string;
+      data: PayPalData;
+    };
 
 export interface ClientToken {
   client_token: string;
 }
 
-export interface PaymentMethodData {
+export interface PaymentMethodPayload {
   type: 'credit_card' | 'payment_method_nonce';
   data: SaveCreditCardData | { nonce: string };
   is_default: boolean;