+++ packages/manager/src/featureFlags.ts
@@ -38,10 +38,11 @@ export interface Flags {
   referralBannerText: ReferralBannerText;
   blockStorageAvailability: boolean;
   imagesPriceInfo: boolean;
-  productInformationBanners: ProductInformationBannerFlag[];
   apiMaintenance: APIMaintenance;
+  productInformationBanners: ProductInformationBannerFlag[];
+  lkeHighAvailability: boolean;
   autoscaler: boolean;
-  deprecateLinodeDomainBanner: boolean;
+  kubernetesDashboardAvailability: boolean;
 }
 
 type PromotionalOfferFeature =