+++ packages/manager/src/features/Billing/BillingPanels/PaymentInfoPanel/AddPaymentMethodDrawer/AddPaymentMethodDrawer.tsx
@@ -12,6 +12,8 @@ import AddCreditCardForm from './AddCreditCardForm';
 import useFlags from 'src/hooks/useFlags';
 import Notice from 'src/components/Notice';
 import { MAXIMUM_PAYMENT_METHODS } from 'src/constants';
+import { PayPalChip } from '../PayPalChip';
+import PayPalErrorBoundary from '../PayPalErrorBoundary';
 
 interface Props {
   open: boolean;
@@ -26,6 +28,7 @@ const useStyles = makeStyles((theme: Theme) => ({
   },
   root: {
     marginTop: 4,
+    marginBottom: 4,
   },
   progress: {
     marginBottom: 18,
@@ -80,6 +83,8 @@ export const AddPaymentMethodDrawer: React.FC<Props> = (props) => {
     'google_pay'
   );
 
+  const isPayPalEnabled = flags.additionalPaymentMethods?.includes('paypal');
+
   const disabled = isProcessing || hasMaxPaymentMethods;
 
   return (
@@ -119,14 +124,39 @@ export const AddPaymentMethodDrawer: React.FC<Props> = (props) => {
           </Grid>
         </>
       ) : null}
+      {isPayPalEnabled ? (
+        <>
+          <Divider />
+          <Grid className={classes.root} container>
+            <Grid item xs={8} md={9}>
+              <Typography variant="h3">PayPal</Typography>
+              <Typography>
+                You’ll be taken to PayPal to complete sign up.
+              </Typography>
+            </Grid>
+            <Grid
+              container
+              item
+              xs={4}
+              md={3}
+              justify="flex-end"
+              alignContent="center"
+            >
+              <PayPalErrorBoundary>
+                <PayPalChip
+                  onClose={onClose}
+                  setProcessing={setIsProcessing}
+                  disabled={disabled}
+                />
+              </PayPalErrorBoundary>
+            </Grid>
+          </Grid>
+        </>
+      ) : null}
       <>
-        <Divider spacingBottom={16} spacingTop={16} />
+        <Divider spacingBottom={16} />
         <Typography variant="h3">Credit Card</Typography>
-        <AddCreditCardForm
-          hasNoPaymentMethods={paymentMethods?.length === 0}
-          disabled={disabled}
-          onClose={onClose}
-        />
+        <AddCreditCardForm disabled={disabled} onClose={onClose} />
       </>
     </Drawer>
   );